telegram-ai-chatbot/requirements.txt
aiogram>=3.22.0
openai>=2.8.0
python-dotenv>=1.2.0
aiosqlite>=0.21.0

